fix /angular circular dependencies
* fix #2955
* moves ng types into it's own plain file, fix order of include/export to reflect heirachy
Note: gridComp depends on gridItem but not other way around
