+++ angular/projects/lib/src/index.ts
@@ -2,7 +2,8 @@
  * Public API Surface of gridstack-angular
  */
 
+export * from './lib/types';
+export * from './lib/base-widget';
 export * from './lib/gridstack-item.component';
 export * from './lib/gridstack.component';
-export * from './lib/base-widget';
 export * from './lib/gridstack.module';
